% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_synthetic_control.R
\name{compute_synthetic_control}
\alias{compute_synthetic_control}
\title{compute_synthetic_control}
\usage{
compute_synthetic_control(
  prepared_dataset,
  unit_of_interest,
  serie_of_interest,
  col_time,
  max_time_unit_of_interest
)
}
\arguments{
\item{prepared_dataset}{Output from prepare_dataset().}

\item{unit_of_interest}{Value of the col_unit_name that is of interest.}

\item{serie_of_interest}{Column name os the serie to be projected.}

\item{col_time}{String with the column name of the time column.}

\item{max_time_unit_of_interest}{Outout from intern_get_max_time_unit_of_interest().}
}
\value{
List with (i) Synth::dataprep() output and (ii) Synth::Synth() output.
}
\description{
Internal function. Please refer to run_synthetic_forecast documentation.
}
\details{
Compute the synthetic control 9wraps Synth package).
}
