% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_example}
\alias{df_example}
\title{Dataset with a y series to be forecasted, and its cumulated value.}
\format{
A dataframe with 1275 rows e 32 variaveis:
\describe{
  \item{unit}{Unit identification}
  \item{time_period}{Time period, ascending integers.}
  \item{x1-x28}{Variables.}
}
}
\usage{
df_example
}
\description{
A generic dataset with a Y value and X predictors.
}
\keyword{datasets}
