% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_elegile_units.R
\name{intern_elegile_units}
\alias{intern_elegile_units}
\title{intern_elegile_units}
\usage{
intern_elegile_units(
  df,
  col_unit_name,
  col_time,
  max_time_unit_of_interest,
  periods_to_forecast
)
}
\arguments{
\item{df}{Main DataFrame.}

\item{col_unit_name}{String with column name of the column with the units names.}

\item{col_time}{String with the column name of the time column.}

\item{max_time_unit_of_interest}{Outout from intern_get_max_time_unit_of_interest().}

\item{periods_to_forecast}{(Integer) Number of periods to forecast.}
}
\value{
DataFrame with the columns: (i) col_unit_name and (ii) manter (bool)
}
\description{
Internal function. Please refer to run_synthetic_forecast documentation.
}
\details{
Selects the elegible units to build the synthetic control: Rule the elegible units are the units that
have at least max_time_unit_of_interest + periods_to_forecast time periods.
}
