% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_get_max_time_unit_of_interest.R
\name{intern_get_max_time_unit_of_interest}
\alias{intern_get_max_time_unit_of_interest}
\title{intern_get_max_time_unit_of_interest}
\usage{
intern_get_max_time_unit_of_interest(
  df,
  col_unit_name,
  unit_of_interest,
  col_time
)
}
\arguments{
\item{df}{Main DataFrame.}

\item{col_unit_name}{String with column name of the column with the units names.}

\item{unit_of_interest}{Value of the col_unit_name that is of interest.}

\item{col_time}{String with the column name of the time column.}
}
\value{
Same type as col_time, max value.
}
\description{
Internal function. Please refer to run_synthetic_forecast documentation.
}
\details{
Intern function to compute the max time period of the unit of interest.
}
