% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_synthetic_forecast.R
\name{run_synthetic_forecast}
\alias{run_synthetic_forecast}
\title{run_synthetic_forecast}
\usage{
run_synthetic_forecast(
  df,
  col_unit_name,
  unit_of_interest,
  col_time,
  periods_to_forecast,
  serie_of_interest
)
}
\arguments{
\item{df}{Main DataFrame.}

\item{col_unit_name}{String with column name of the column with the units names.}

\item{unit_of_interest}{Value of the col_unit_name that is of interest.}

\item{col_time}{String with the column name of the time column.}

\item{periods_to_forecast}{(Integer) Number of periods to forecast.}

\item{serie_of_interest}{Column name os the serie to be projected.}
}
\value{
List with results table.
}
\description{
Executes all the other package functions in order to have a list with the results table.
}
\examples{
\donttest{
synthetic_forecast <- run_synthetic_forecast(
df = df_example,
col_unit_name = 'unit',
col_time='time_period',
periods_to_forecast=12,
unit_of_interest = '30',
serie_of_interest = 'x1'
)
}

}
