% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{prepare_dataset}
\usage{
prepare_dataset(
  df,
  df_elegible_units,
  col_unit_name,
  col_time,
  unit_of_interest,
  max_time_unit_of_interest
)
}
\arguments{
\item{df}{Main DataFrame.}

\item{df_elegible_units}{output from intern_elegile_units().}

\item{col_unit_name}{String with column name of the column with the units names.}

\item{col_time}{String with the column name of the time column.}

\item{unit_of_interest}{Value of the col_unit_name that is of interest.}

\item{max_time_unit_of_interest}{Outout from intern_get_max_time_unit_of_interest().}
}
\value{
A dataset to be inputed in the compute_synthetic_control().
}
\description{
Internal function. Please refer to run_synthetic_forecast documentation.
}
